Don't perform unnecessary installs during `paket add`

+++ src/Paket.Core/AddProcess.fs
@@ -13,22 +13,26 @@ let private addToProject (project : ProjectFile) package =
         .AddNuGetReference(package)
         .Save()
 
-let private add addToProjects dependenciesFileName package version force hard installAfter =
+let private add installToProjects addToProjectsF dependenciesFileName package version force hard installAfter =
     let existingDependenciesFile = DependenciesFile.ReadFromFile(dependenciesFileName)
-    let dependenciesFile =
-        existingDependenciesFile
-            .Add(package,version)
+    let (PackageName name) = package
+    if (not installToProjects) && existingDependenciesFile.HasPackage package then 
+        traceWarnfn "%s contains package %s already." dependenciesFileName name
+    else
+        let dependenciesFile =
+            existingDependenciesFile
+                .Add(package,version)
     
-    let lockFile = UpdateProcess.SelectiveUpdate(dependenciesFile,Some(NormalizedPackageName package),force)
-    let projects = seq { for p in ProjectFile.FindAllProjects(Path.GetDirectoryName lockFile.FileName) -> p } // lazy sequence in case no project install required
+        let lockFile = UpdateProcess.SelectiveUpdate(dependenciesFile,Some(NormalizedPackageName package),force)
+        let projects = seq { for p in ProjectFile.FindAllProjects(Path.GetDirectoryName lockFile.FileName) -> p } // lazy sequence in case no project install required
 
-    dependenciesFile.Save()
+        dependenciesFile.Save()
     
-    package |> addToProjects projects
+        package |> addToProjectsF projects
 
-    if installAfter then
-        let sources = dependenciesFile.GetAllPackageSources()
-        InstallProcess.Install(sources, force, hard, false, lockFile)
+        if installAfter then
+            let sources = dependenciesFile.GetAllPackageSources()
+            InstallProcess.Install(sources, force, hard, false, lockFile)
 
 // add a package with the option to add it to a specified project
 let AddToProject(dependenciesFileName, package, version, force, hard, projectName, installAfter) =
@@ -45,7 +49,7 @@ let AddToProject(dependenciesFileName, package, version, force, hard, projectNam
         | None ->
             traceErrorfn "Could not install package in specified project %s. Project not found" projectName
 
-    add addToSpecifiedProject dependenciesFileName package version force hard installAfter
+    add true addToSpecifiedProject dependenciesFileName package version force hard installAfter
     
 // add a package with the option to interactively add it to multiple projects
 let Add(dependenciesFileName, package, version, force, hard, interactive, installAfter) =
@@ -56,4 +60,4 @@ let Add(dependenciesFileName, package, version, force, hard, interactive, instal
                 if package |> notInstalled project && Utils.askYesNo(sprintf "  Install to %s?" project.Name) then
                     package |> addToProject project
     
-    add addToProjects dependenciesFileName package version force hard installAfter
+    add interactive addToProjects dependenciesFileName package version force hard installAfter
